REFACTOR: Split filtering files and put all filtering files in one folder
